Move pkg/vclib and pkg/config under pkg/common
