+++ pkg/cloudprovider/vsphere/types.go
@@ -21,10 +21,10 @@ import (
 
 	"k8s.io/api/core/v1"
 	clientv1 "k8s.io/client-go/listers/core/v1"
-	"k8s.io/cloud-provider-vsphere/pkg/vclib"
+	"k8s.io/cloud-provider-vsphere/pkg/common/vclib"
 	"k8s.io/kubernetes/pkg/cloudprovider"
 
-	vcfg "k8s.io/cloud-provider-vsphere/pkg/config"
+	vcfg "k8s.io/cloud-provider-vsphere/pkg/common/config"
 )
 
 // GRPCServer interface
