+++ pkg/common/vclib/custom_errors.go
