+++ pkg/common/vclib/fixtures/fixtures.go
