+++ pkg/common/vclib/utils.go
