+++ pkg/common/vclib/virtualmachine.go
