+++ pkg/common/vclib/virtualmachine_test.go
